Extract NewInstanceValidator from ConfigValidator
